Refactor: repository 추상체 위치 변경
