--- core-repository/src/main/java/com/hmoa/core_repository/CommunityComment/CommunityCommentRepository.kt
@@ -1,20 +1,0 @@
-package com.hmoa.core_repository.CommunityComment
-
-import com.hmoa.core_model.request.CommunityCommentDefaultRequestDto
-import com.hmoa.core_model.response.CommunityCommentAllResponseDto
-import com.hmoa.core_model.response.CommunityCommentDefaultResponseDto
-import com.hmoa.core_model.response.DataResponseDto
-
-interface CommunityCommentRepository {
-    suspend fun putCommunityComment(
-        commentId: Int,
-        dto: CommunityCommentDefaultRequestDto
-    ): CommunityCommentDefaultResponseDto
-
-    suspend fun deleteCommunityComment(commentId: Int): DataResponseDto<Any>
-    suspend fun getCommunityComments(commentId: Int, page: String): CommunityCommentAllResponseDto
-    suspend fun postCommunityComment(
-        commentId: Int,
-        dto: CommunityCommentDefaultRequestDto
-    ): CommunityCommentDefaultResponseDto
-}
