--- core-repository/src/main/java/com/hmoa/core_repository/Fcm/FcmRepository.kt
@@ -1,10 +1,0 @@
-package com.hmoa.core_repository.Fcm
-
-import com.hmoa.core_model.request.FCMTokenSaveRequestDto
-import com.hmoa.core_model.response.DataResponseDto
-
-interface FcmRepository {
-    suspend fun deleteFcmToken() : DataResponseDto<Any>
-
-    suspend fun saveFcmToken(fcmTokenSaveRequest : FCMTokenSaveRequestDto) : DataResponseDto<Any>
-}
