--- core-repository/src/main/java/com/hmoa/core_repository/Perfumer/PerfumerRepository.kt
@@ -1,8 +1,0 @@
-package com.hmoa.core_repository.Perfumer
-
-import com.hmoa.core_model.response.DataResponseDto
-
-interface PerfumerRepository {
-    suspend fun getPerfumers(pageNum: String): DataResponseDto<Any>
-    suspend fun getPerfumer(perfumerId: Int): DataResponseDto<Any>
-}
