+++ core-repository/src/main/java/com/hmoa/core_repository/RepositoryModule.kt
@@ -1,33 +1,6 @@
 package com.hmoa.core_repository
 
-import com.hmoa.core_repository.Admin.AdminRepository
-import com.hmoa.core_repository.Admin.AdminRepositoryImpl
-import com.hmoa.core_repository.Brand.BrandRepository
-import com.hmoa.core_repository.Brand.BrandRepositoryImpl
-import com.hmoa.core_repository.BrandHPedia.BrandHPediaRepository
-import com.hmoa.core_repository.BrandHPedia.BrandHPediaRepositoryImpl
-import com.hmoa.core_repository.Community.CommunityRepository
-import com.hmoa.core_repository.Community.CommunityRepositoryImpl
-import com.hmoa.core_repository.CommunityComment.CommunityCommentRepository
-import com.hmoa.core_repository.CommunityComment.CommunityCommentRepositoryImpl
-import com.hmoa.core_repository.Fcm.FcmRepository
-import com.hmoa.core_repository.Fcm.FcmRepositoryImpl
-import com.hmoa.core_repository.Login.LoginRepository
-import com.hmoa.core_repository.Login.LoginRepositoryImpl
-import com.hmoa.core_repository.Main.MainRepository
-import com.hmoa.core_repository.Main.MainRepositoryImpl
-import com.hmoa.core_repository.Member.MemberRepository
-import com.hmoa.core_repository.Member.MemberRepositoryImpl
-import com.hmoa.core_repository.Note.NoteRepository
-import com.hmoa.core_repository.Note.NoteRepositoryImpl
-import com.hmoa.core_repository.Perfume.PerfumeRepository
-import com.hmoa.core_repository.Perfume.PerfumeRepositoryImpl
-import com.hmoa.core_repository.PerfumeComment.PerfumeCommentRepository
-import com.hmoa.core_repository.PerfumeComment.PerfumeCommentRepositoryImpl
-import com.hmoa.core_repository.Perfumer.PerfumerRepository
-import com.hmoa.core_repository.Perfumer.PerfumerRepositoryImpl
-import com.hmoa.core_repository.Search.SearchRepository
-import com.hmoa.core_repository.Search.SearchRepositoryImpl
+import com.hmoa.core_domain.repository.*
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
